@@ -7,19 +7,22 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class Instance {
-	public final String PROJECT_MARK = "try-";
-	
-	public List<ParsedNews> generate(String lineOfData){
-		List<ParsedNews> results = new ArrayList<ParsedNews>();
+
+	public List<StoryPojo> generate(String lineOfData){
+		List<StoryPojo> results = new ArrayList<StoryPojo>();
 		List<String> parts = Arrays.asList(lineOfData.split("\\\n"));
 		
 		for(String part : parts) {
-			ParsedNews result = new ParsedNews();
+			StoryPojo result = new StoryPojo();
 			String[] fragments = part.split("\\|");
-			result.setTitle(PROJECT_MARK+fragments[0]);
-			result.setBody(PROJECT_MARK+fragments[1]);
-			result.setLink(PROJECT_MARK+fragments[2]);
-			result.setTags(Stream.of(fragments[3].split(", ")).map(e -> PROJECT_MARK + e).collect(Collectors.toList()));
+			result.setId(fragments[0]);
+			result.setTitle(fragments[1]);
+			result.setUrl(fragments[2]);
+			result.setSite(fragments[3]);
+			result.setTime(Long.valueOf(fragments[4]));
+			result.setFavicon_url(fragments[5]);
+			result.setTags(Stream.of(fragments[6].split(", ")).collect(Collectors.toList()));
+			result.setSimilar_stories(Stream.of(fragments[7].split(", ")).collect(Collectors.toList()));
 			results.add(result);
 		}
 		return results;
