@@ -5,23 +5,23 @@ import java.io.IOException;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 
-import com.example.sharov.anatoliy.flink.protobuf.NewsProtos.News;
+import com.example.sharov.anatoliy.flink.protobuf.StoryProtos.Story;;
 
 @SuppressWarnings("serial")
-public class CustomProtobufDeserializer implements DeserializationSchema<News>{
+public class CustomProtobufDeserializer implements DeserializationSchema<Story>{
 
 	@Override
-	public TypeInformation<News> getProducedType() {
-		return TypeInformation.of(News.class);
+	public TypeInformation<Story> getProducedType() {
+		return TypeInformation.of(Story.class);
 	}
 
 	@Override
-	public News deserialize(byte[] message) throws IOException {
-		return News.parseFrom(message);
+	public Story deserialize(byte[] message) throws IOException {
+		return Story.parseFrom(message);
 	}
 
 	@Override
-	public boolean isEndOfStream(News nextElement) {
+	public boolean isEndOfStream(Story nextElement) {
 		return false;
 	}
 
