@@ -1,5 +1,5 @@
 package com.example.sharov.anatoliy.producertickertick;
 
-public class NewsPojo {
+public class TickertickController {
 
 }
